backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
email-validator==2.1.0
pandas==2.1.4
urllib3==2.1.0